+++ CHANGELOG.md
@@ -53,8 +53,10 @@ high state of flux, you're at risk of it changing without notice.
     - make `zip` data-last (@gcanti)
   - `ReadonlyRecord`
     - make `isSubrecord` data-last (@gcanti)
+    - make `lookup` data-last (@gcanti)
   - `Record`
     - make `isSubrecord` data-last (@gcanti)
+    - make `lookup` data-last (@gcanti)
   - `Semigroup`
     - make `fold` data-last (@gcanti)